Added a function in constroctor for Card. Also overloaded the () operator.
Added a 3x3 array of char in Card.cpp that we will use to print out the board. Also overloaded the () operator as required to access a row of the card as a string. Added another 3x3 array as a const static member variable to diplay the back of card.